UPD: 23, 24

+++ posts/views.py
@@ -111,10 +111,10 @@ def edit_note_view(request: WSGIRequest, note_uuid):
         note.title = request.POST["title"]
         note.content = request.POST["content"]
         note.mod_time = timezone.now()
-        if note.image:
-            note.image.delete()
-            note.image = request.FILES["noteImage"]
-            note.save()
+        if request.FILES.get("noteImage"):
+            if note.image:
+                note.image.delete()
+            note.image = request.FILES.get("noteImage")
         note.save()
         return HttpResponseRedirect(reverse('show-note', args=[note.uuid]))
